+++ lib/provider.dart
@@ -1,7 +1,7 @@
 library provider;
 
+export 'src/async_provider.dart';
 export 'src/consumer.dart';
 export 'src/delegate_widget.dart';
 export 'src/listenable_provider.dart';
 export 'src/provider.dart';
-export 'src/stream_provider.dart';