+++ lib/src/async_provider.dart
@@ -9,7 +9,7 @@ typedef ErrorBuilder<T> = T Function(BuildContext context, Object error);
 /// Listens to a [Stream<T>] and exposes [T] to its descendants.
 ///
 /// It is considered an error to pass a stream that can emit errors without providing
-/// an [catchError] method.
+/// a [catchError] method.
 ///
 /// {@template provider.streamprovider.initialdata}
 /// [initialData] determines the value exposed until the [Stream] emits a value.
@@ -74,8 +74,8 @@ class StreamProvider<T> extends ValueDelegateWidget<Stream<T>>
   /// {@macro provider.streamprovider.initialdata}
   final T initialData;
 
-  /// The widget that is below the current [StreamProvider] widget in the
-  /// tree.
+  /// The widget that is below the current [StreamProvider] widget in the tree.
+  ///
   /// {@macro flutter.widgets.child}
   final Widget child;
 
@@ -109,7 +109,7 @@ class StreamProvider<T> extends ValueDelegateWidget<Stream<T>>
       stream: delegate.value,
       initialData: initialData,
       builder: (_, snapshot) {
-        return Provider<T>.value(
+        return InheritedProvider<T>(
           value: _getValue(snapshot, context, catchError),
           child: child,
           updateShouldNotify: updateShouldNotify,
@@ -156,5 +156,111 @@ class _StreamControllerBuilderDelegate<T>
   @override
   void dispose() {
     controller?.close();
+  }
+}
+
+/// Listens to a [Future<T>] and exposes [T] to its descendants.
+///
+/// It is considered an error to pass a future that can emit errors without providing
+/// a [catchError] method.
+///
+/// {@macro provider.updateshouldnotify}
+///
+/// See also:
+///   * [Future]
+class FutureProvider<T> extends ValueDelegateWidget<Future<T>>
+    implements SingleChildCloneableWidget {
+  /// Creates a [Future] from [builder] and subscribes to it.
+  ///
+  /// [builder] must not be `null`.
+  FutureProvider({
+    Key key,
+    @required ValueBuilder<Future<T>> builder,
+    T initialData,
+    ErrorBuilder<T> catchError,
+    UpdateShouldNotify<T> updateShouldNotify,
+    Widget child,
+  }) : this._(
+          key: key,
+          initialData: initialData,
+          catchError: catchError,
+          updateShouldNotify: updateShouldNotify,
+          delegate: BuilderAdaptiveDelegate(builder),
+          child: child,
+        );
+
+  /// Listens to [future] and expose it to all of [FutureProvider] descendants.
+  FutureProvider.value({
+    Key key,
+    @required Future<T> future,
+    T initialData,
+    ErrorBuilder<T> catchError,
+    UpdateShouldNotify<T> updateShouldNotify,
+    Widget child,
+  }) : this._(
+          key: key,
+          initialData: initialData,
+          catchError: catchError,
+          updateShouldNotify: updateShouldNotify,
+          delegate: SingleValueDelegate(future),
+          child: child,
+        );
+
+  FutureProvider._({
+    Key key,
+    ValueAdaptiveDelegate<Future<T>> delegate,
+    this.initialData,
+    this.catchError,
+    this.updateShouldNotify,
+    this.child,
+  }) : super(key: key, delegate: delegate);
+
+  /// [initialData] determines the value exposed until the [Future] completes.
+  ///
+  /// If omitted, defaults to `null`.
+  final T initialData;
+
+  /// The widget that is below the current [FutureProvider] widget in the tree.
+  ///
+  /// {@macro flutter.widgets.child}
+  final Widget child;
+
+  /// Optional function used if the [Future] emits an error.
+  ///
+  /// [catchError] will be called with the emitted error and
+  /// is expected to return a fallback value without throwing.
+  ///
+  /// The returned value will then be exposed to the descendants of [FutureProvider]
+  /// like any valid value.
+  final ErrorBuilder<T> catchError;
+
+  /// {@macro provider.updateshouldnotify}
+  final UpdateShouldNotify<T> updateShouldNotify;
+
+  @override
+  FutureProvider<T> cloneWithChild(Widget child) {
+    return FutureProvider._(
+      key: key,
+      delegate: delegate,
+      updateShouldNotify: updateShouldNotify,
+      initialData: initialData,
+      catchError: catchError,
+      child: child,
+    );
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return FutureBuilder<T>(
+      future: delegate.value,
+      initialData: initialData,
+      builder: (_, snapshot) {
+        return InheritedProvider<T>(
+          value: _getValue(snapshot, context, catchError),
+          updateShouldNotify: updateShouldNotify,
+          child: child,
+        );
+      },
+    );
   }
 }