+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 3.0.0
 
+- Added `FutureProvider`, which takes a future and updates dependents when the future completes.
 - Providers can no longer be instantiated using `const` constructors.
 - Added `DelegateWidget` and a few related classes to help building custom providers.
 - Exposed the internal generic `InheritedWidget` to help building custom providers.